llm-research-agent/requirements.txt
openai==0.28.1
google-genai==1.23.0
redis==4.6.0
fastapi
uvicorn
langgraph
httpx
pytest
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
prometheus_client

